Split mutable and immutable into own files, update usage
