+++ .verb.md
@@ -10,3 +10,11 @@ defaults(obj, {a: 'bbb', d: 'c'});
 console.log(obj);
 //=> {a: 'c', d: 'c'}
 ```
+
+Or immutable defaulting:
+```js
+var defaults = require('{%= name %}/immutable');
+var obj = {a: 'c'};
+var defaulted = defaults(obj, {a: 'bbb', d: 'c'});
+console.log(obj !== defaulted);
+//=> true
